+++ vscode/src/jsMain/kotlin/vscode/InlineValueContext.kt
@@ -3,12 +3,11 @@ package vscode
 /**
  * A value-object that contains contextual information when requesting inline values from a InlineValuesProvider.
  */
-
 external interface InlineValueContext {
   /**
    * The stack frame (as a DAP Id) where the execution has stopped.
    */
-  val frameId: Double
+  val frameId: Int
 
   /**
    * The document range where execution has stopped.
