+++ vscode/src/jsMain/kotlin/vscode/TaskProcessEndEvent.kt
@@ -4,7 +4,6 @@ package vscode
  * An event signaling the end of a process execution
  * triggered through a task
  */
-
 external interface TaskProcessEndEvent {
   /**
    * The task execution for which the process got started.
@@ -14,5 +13,5 @@ external interface TaskProcessEndEvent {
   /**
    * The process's exit code. Will be `undefined` when the task is terminated.
    */
-  val exitCode: Double?
+  val exitCode: Int?
 }
