+++ vscode/src/jsMain/kotlin/vscode/TaskProcessStartEvent.kt
@@ -4,7 +4,6 @@ package vscode
  * An event signaling the start of a process execution
  * triggered through a task
  */
-
 external interface TaskProcessStartEvent {
   /**
    * The task execution for which the process got started.
@@ -14,5 +13,5 @@ external interface TaskProcessStartEvent {
   /**
    * The underlying process id.
    */
-  val processId: Double
+  val processId: Int
 }
