+++ vscode/src/jsMain/kotlin/vscode/TerminalDimensions.kt
@@ -3,15 +3,14 @@ package vscode
 /**
  * Represents the dimensions of a terminal.
  */
-
 external interface TerminalDimensions {
   /**
    * The number of columns in the terminal.
    */
-  val columns: Double
+  val columns: Int
 
   /**
    * The number of rows in the terminal.
    */
-  val rows: Double
+  val rows: Int
 }
