+++ vscode/src/jsMain/kotlin/vscode/TerminalExitStatus.kt
@@ -3,7 +3,6 @@ package vscode
 /**
  * Represents how a terminal exited.
  */
-
 external interface TerminalExitStatus {
   /**
    * The exit code that a terminal exited with, it can have the following values:
@@ -12,7 +11,7 @@ external interface TerminalExitStatus {
    * - `undefined`: the user forcibly closed the terminal or a custom execution exited
    *   without providing an exit code.
    */
-  val code: Double?
+  val code: Int?
 
   /**
    * The reason that triggered the exit of a terminal.
