+++ vscode/src/jsMain/kotlin/vscode/TerminalLink.kt
@@ -12,11 +12,11 @@ external class TerminalLink {
    * @param length The length of the link on [TerminalLinkContext.line].
    * @param tooltip The tooltip text when you hover over this link.
    *
-   * If a tooltip is provided, is will be displayed in a string that includes instructions on
-   * how to trigger the link, such as `{0} (ctrl + click)`. The specific instructions vary
-   * depending on OS, user settings, and localization.
+   *   If a tooltip is provided, is will be displayed in a string that includes instructions on
+   *   how to trigger the link, such as `{0} (ctrl + click)`. The specific instructions vary
+   *   depending on OS, user settings, and localization.
    */
-  constructor(startIndex: Double, length: Double, tooltip: String = definedExternally)
+  constructor(startIndex: Int, length: Int, tooltip: String = definedExternally)
 
   /**
    * The start index of the link on [TerminalLinkContext.line].
