+++ vscode/src/jsMain/kotlin/vscode/TextDocumentContentChangeEvent.kt
@@ -1,26 +1,25 @@
 package vscode
 
 /**
- * An event describing an individual change in the text of a [TextDocument document].
+ * An event describing an individual change in the text of a [TextDocument].
  */
-
 external interface TextDocumentContentChangeEvent {
-/**
+  /**
    * The range that got replaced.
    */
   val range: Range
 
-/**
+  /**
    * The offset of the range that got replaced.
    */
-  val rangeOffset: Double
+  val rangeOffset: Int
 
-/**
+  /**
    * The length of the range that got replaced.
    */
-  val rangeLength: Double
+  val rangeLength: Int
 
-/**
+  /**
    * The new text for the range.
    */
   val text: String
