+++ vscode/src/jsMain/kotlin/vscode/WebviewPortMapping.kt
@@ -7,10 +7,10 @@ external interface WebviewPortMapping {
   /**
    * Localhost port to remap inside the webview.
    */
-  val webviewPort: Double
+  val webviewPort: Int
 
   /**
    * Destination port. The `webviewPort` is resolved to this port.
    */
-  val extensionHostPort: Double
+  val extensionHostPort: Int
 }
